Time subsystem overhaul (test: ./timer_test.exp)
The timing functionality currently in use in xv6 is limited and often
incorrect in several ways:

 1. It lacks proper initialization routines for its timers, relying on
    the host OS or by simple guessing. The frequency of the TSC is
    determined during compilation by parsing '/proc/cpuinfo', which is
    incorrect, since modern processors employ dynamic frequencies that
    can change drastically, while the TSC frequency is constant.

 2. The frequency of the interrupt timer is just set to a constant value
    without any calibration, which results in a significant drift of the
    tick counter compared to the TSC.

 3. There is no precise definition of a jiffy (a tick) and its relation
    to the timers, even though it's a core concept of an OS.

 4. The time calculation functions are implemented in a very inefficient
    way, making them useless for high-precision profiling. In addition,
    the data type whey use (long long) is very error-prone and is often
    used without proper casting.

This patch attempts to significantly improve the time subsystem in the
following ways:

 1. The kernel jiffy and other intervals are precisely defined in a
    single place.

 2. The timers are calibrated using a well-known time source (the PIT).
    Their values are kept synced, so the processor cycle count and the
    interrupt tick count are within the same time-frame.

 3. High precision measurement functions are implemented. When
    millisecond precision is required, time can be measured since boot.
    For microsecond precision, the function is split in two: making a
    timestamp, and measuring the time from that timestamp. This is done
    to avoid integer overflows, eliminating the need for 'long long' and
    its complex handling.

 4. Similarly, the sleep functions split the waiting time to jiffies,
    which can be slept upon, and microseconds, which should be
    busy-waited.

 5. Wakeups are performed only on the relevant ticks, since they require
    much overhead - which is critical for interrupts.

The patch is split into the following files:

 - clock.h: The definitions of a jiffy, millisecond, microsecond, and
   various functions for their measurements.

 - clock.c: Handles the majority of time-related functions, including
   calibration of timers and handling time interrupts.

 - clockasm.S: The hardware interface of the timers. It includes the
   routines for measuring the intervals using the PIT - which should be
   extremely precise, and functions that require reading the TSC and
   performing 64-bit divisions - so they remain efficient and abstracted
   from the C code.

Limitations:

 - The expected precision is +/- 5sec/hour.

 - It will take about 10 seconds to stabilize the interrupt frequency
   after boot.

 - The time since boot can be measured in milliseconds for 48 days,
   after which the counter overflows.

 - The time from a timestamp in microseconds can be measured for 70
   minutes, after which it overflows.

 - The assembly calibration routines must not be paused for debugging,
   or the results will be invalid. The C side of the calibration can be
   paused.

 - When running with kvm emulation (not the default), the timer
   interrupt frequency will increase after pausing the VM, since the TSC
   keeps running in paused state. Pausing will reduce the interrupt
   interval by 10% for a while, but it won't affect time measurement.

Tests:

 - A test was added to the main test script (runtests.exp)

 - A (temporary) dedicated test script was added - timer_test.exp.
   *** PLEASE RUN IT AND SEE IF IT WORKS! ***

+++ cgroup.h
@@ -73,7 +73,7 @@ struct cgroup
 
     unsigned int max_mem; /*The maximum memory allowed for a group to use.*/
 
-    unsigned long long cpu_time;
+    unsigned int cpu_time;
     unsigned int cpu_period_time;
     unsigned int cpu_percent;
     unsigned int cpu_account_period;
